varsity_plug/requirements.txt
redis==5.0.8 
django-redis==5.4.0 
